refactor: Social패키지 재정리 및 통합 #6
