--- src/main/java/cmc/delta/domain/auth/api/dto/KakaoLoginData.java
@@ -1,7 +1,0 @@
-package cmc.delta.domain.auth.api.dto;
-
-public record KakaoLoginData(String email, String nickname, boolean isNewUser) {
-	public static KakaoLoginData of(String email, String nickname, boolean isNewUser) {
-		return new KakaoLoginData(email, nickname, isNewUser);
-	}
-}
