--- src/main/java/cmc/delta/domain/auth/api/dto/KakaoLoginRequest.java
@@ -1,7 +1,0 @@
-package cmc.delta.domain.auth.api.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record KakaoLoginRequest(@NotBlank
-String code) {
-}
