+++ src/main/java/cmc/delta/domain/auth/api/dto/request/SocialLoginRequest.java
@@ -1,0 +1,7 @@
+package cmc.delta.domain.auth.api.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record SocialLoginRequest(@NotBlank
+String code) {
+}
