+++ src/main/java/cmc/delta/domain/auth/api/dto/response/SocialAccountData.java
@@ -1,0 +1,4 @@
+package cmc.delta.domain.auth.api.dto.response;
+
+public class SocialAccountData {
+}
