--- src/main/java/cmc/delta/domain/auth/application/SocialAuthService.java
@@ -1,3 +1,0 @@
-package cmc.delta.domain.auth.application;
-
-public class SocialAuthService {}
