--- src/main/java/cmc/delta/domain/auth/infrastructure/oauth/AppleOAuthClient.java
@@ -1,3 +1,0 @@
-package cmc.delta.domain.auth.infrastructure.oauth;
-
-public class AppleOAuthClient {}
