+++ src/main/java/cmc/delta/domain/auth/infrastructure/oauth/apple/AppleOAuthClient.java
@@ -1,0 +1,3 @@
+package cmc.delta.domain.auth.infrastructure.oauth.apple;
+
+public class AppleOAuthClient {}
