+++ src/main/java/cmc/delta/domain/auth/infrastructure/oauth/apple/AppleOAuthConfig.java
@@ -1,0 +1,4 @@
+package cmc.delta.domain.auth.infrastructure.oauth.apple;
+
+public class AppleOAuthConfig {
+}
