+++ src/main/java/cmc/delta/domain/auth/infrastructure/oauth/client/OAuthHttpClient.java
@@ -1,0 +1,4 @@
+package cmc.delta.domain.auth.infrastructure.oauth.client;
+
+public class OAuthHttpClient {
+}
