+++ src/main/java/cmc/delta/domain/auth/infrastructure/oauth/token/redis/RedisAccessBlacklistStore.java
@@ -1,4 +1,4 @@
-package cmc.delta.domain.auth.infrastructure.redis;
+package cmc.delta.domain.auth.infrastructure.oauth.token.redis;
 
 import cmc.delta.domain.auth.application.port.AccessBlacklistStore;
 import java.time.Duration;
