--- src/main/java/cmc/delta/domain/auth/infrastructure/persistence/SocialAccountJpaRepository.java
@@ -1,3 +1,0 @@
-package cmc.delta.domain.auth.infrastructure.persistence;
-
-public class SocialAccountJpaRepository {}
