+++ src/main/java/cmc/delta/domain/auth/persistence/SocialAccountJpaRepository.java
@@ -1,0 +1,3 @@
+package cmc.delta.domain.auth.persistence;
+
+public class SocialAccountJpaRepository {}
